SEQNG-118 Changes to appease scalastyle
