+++ modules/edu.gemini.seqexec.engine/src/main/scala/edu/gemini/seqexec/engine/Event.scala
@@ -5,7 +5,6 @@ package edu.gemini.seqexec.engine
 
 import scalaz._
 import scalaz.syntax.std.option._
-import scalaz.std.string._
 
 import edu.gemini.seqexec.model.Model.{CloudCover, Conditions, ImageQuality, SkyBackground, WaterVapor, Operator, Observer}
 import edu.gemini.seqexec.model.UserDetails
@@ -19,61 +18,6 @@ import scalaz.stream.Process
 sealed trait Event
 final case class EventUser(ue: UserEvent) extends Event
 final case class EventSystem(se: SystemEvent) extends Event
-
-/**
-  * Events generated by the user.
-  */
-sealed trait UserEvent {
-  def user: Option[UserDetails]
-  def username: String = ~user.map(_.username)
-}
-
-final case class Start(id: Sequence.Id, user: Option[UserDetails]) extends UserEvent
-final case class Pause(id: Sequence.Id, user: Option[UserDetails]) extends UserEvent
-final case class CancelPause(id: Sequence.Id, user: Option[UserDetails]) extends UserEvent
-final case class Load(id: Sequence.Id, sequence: Sequence[Action \/ Result]) extends UserEvent {
-  val user: Option[UserDetails] = None
-}
-final case class Unload(id: Sequence.Id) extends UserEvent {
-  val user: Option[UserDetails] = None
-}
-final case class Breakpoint(id: Sequence.Id, user: Option[UserDetails], step: Step.Id, v: Boolean) extends UserEvent
-final case class SetOperator(name: Operator, user: Option[UserDetails]) extends UserEvent
-final case class SetObserver(id: Sequence.Id, user: Option[UserDetails], name: Observer) extends UserEvent
-final case class SetConditions(conditions: Conditions, user: Option[UserDetails]) extends UserEvent
-final case class SetImageQuality(iq: ImageQuality, user: Option[UserDetails]) extends UserEvent
-final case class SetWaterVapor(wv: WaterVapor, user: Option[UserDetails]) extends UserEvent
-final case class SetSkyBackground(wv: SkyBackground, user: Option[UserDetails]) extends UserEvent
-final case class SetCloudCover(cc: CloudCover, user: Option[UserDetails]) extends UserEvent
-case object Poll extends UserEvent {
-  val user: Option[UserDetails] = None
-}
-// Generic event to put a function in the main Process process, which takes an
-// action depending on the current state
-final case class GetState(f: (Engine.State) => Task[Option[Process[Task, Event]]]) extends UserEvent {
-  val user: Option[UserDetails] = None
-}
-// Calls a user given function in the main Process process to stop an Action.
-// It sets the Sequence to be stopped. The user function is called only if the Sequence is running.
-final case class ActionStop(id: Sequence.Id, f: (Sequence.State) => Option[Process[Task, Event]]) extends UserEvent {
-  val user: Option[UserDetails] = None
-}
-final case class Log(msg: String) extends UserEvent {
-  val user: Option[UserDetails] = None
-}
-
-/**
-  * Events generated internally by the Engine.
-  */
-sealed trait SystemEvent
-final case class Completed[R<:RetVal](id: Sequence.Id, i: Int, r: OK[R]) extends SystemEvent
-final case class PartialResult[R<:PartialVal](id: Sequence.Id, i: Int, r: Partial[R]) extends SystemEvent
-final case class Failed(id: Sequence.Id, i: Int, e: Result.Error) extends SystemEvent
-final case class Busy(id: Sequence.Id) extends SystemEvent
-final case class Executed(id: Sequence.Id) extends SystemEvent
-final case class Executing(id: Sequence.Id) extends SystemEvent
-final case class Finished(id: Sequence.Id) extends SystemEvent
-final object Null extends SystemEvent
 
 object Event {
